#  Speed up EncryptRTP