@Sean-Der 2019-02-14
Fantastic work @kixelated!

Write in place is a poor choice, but I was testing for a realloc (and maybe my assumptions were wrong) but I would hope that people could pass the buffer and we could allocate the tag for them.

Or they could allocation ahead of time (we should add a case to test for this also!) nice little improvement there.

I see three unique inputs though (and would be nice to test for them all)

* New allocation
* realloc for auth tag
* completely in place