@kixelated 2019-02-15
@Sean-Der Thanks for the review and sorry for being so eager to break the API. I'm going to revert the API changes and just improve the performance.