@Sean-Der 2019-02-15
Hey @kixelated 

I think it is awesome you want to break the API actually :) your fresh insight is SUPER valuable, and it would be a huge mistake if I tempered it by accident.

Everything you think is wrong please put on the wiki page, I want to do them all. I just have to balance stability and making everyone happy is my only concern :( and I could be doing a bad job of that so please always push back if I am wrong.
-----

Tell me when you are ready for another review! If there are no API changes that would be fantastic, we can bump in pions/webrtc and keep getting these wins :) really excited about the promise of faster ciphers